Add prompt when init cmd overwrite config file

+++ commands/init_template.py
@@ -8,7 +8,7 @@ class InitTemplateCommand(Command):
     Init new config file base-ed on template. Supported project type: php, ruby, android
 
     init
-        {project_type : This will overwrite any existing config file}
+        {project_type : Supported project type}
     """
 
     def handle(self):
@@ -16,9 +16,20 @@ class InitTemplateCommand(Command):
         if project_type not in self.app.PROJ_TYPES:
             self.line('<error>Invalid project type !</error>')
         else:
-            file_path = os.path.join(self.app.TEMPLATES_DIR, "%s.yml" % project_type)
-            with open(file_path, "r") as fin:
-                with open(self.app.configure_file_name, "w") as fout:
-                    fout.write(fin.read())
-                    self.line('<info>Wrote to file: %s</info>' % self.app.configure_file_name)
+            if os.path.isfile(self.app.configure_file_name):
+                while True:
+                    answer = raw_input("Overwrite file (y/n) ?: ")
+                    if answer in ["y", "n"]:
+                        break
+                    else:
+                        self.line("<comment>Please response 'y' or 'n'</comment>")
+            else:
+                answer = "y"
+
+            if answer == "y":
+                file_path = os.path.join(self.app.TEMPLATES_DIR, "%s.yml" % project_type)
+                with open(file_path, "r") as fin:
+                    with open(self.app.configure_file_name, "w") as fout:
+                        fout.write(fin.read())
+                        self.line('<info>Wrote to file: %s</info>' % self.app.configure_file_name)
         sys.exit(0)
